[surepetcare] changed datatype of key object ids from integer to long to support newly generated ids from Sure Petcare API.

+++ bundles/org.openhab.binding.surepetcare/src/main/java/org/openhab/binding/surepetcare/internal/dto/SurePetcareBaseObject.java
@@ -26,7 +26,7 @@ import org.eclipse.jdt.annotation.NonNullByDefault;
 @NonNullByDefault
 public class SurePetcareBaseObject {
 
-    public Integer id = 0;
+    public Long id = Long.valueOf(0);
     public String version = "";
     public Date createdAt = new Date();
     public Date updatedAt = new Date();
