+++ bundles/org.openhab.binding.surepetcare/src/test/java/org/openhab/binding/surepetcare/internal/data/SurePetcarePetTest.java
@@ -92,7 +92,7 @@ public class SurePetcarePetTest {
         String testReponse = "{\"id\":34675,\"name\":\"Cat\",\"gender\":0,\"date_of_birth\":\"2017-08-01T00:00:00+00:00\",\"weight\":\"3.5\",\"comments\":\"Test Comment\",\"household_id\":87435,\"breed_id\":382,\"photo_id\":23412,\"species_id\":1,\"tag_id\":60456,\"version\":\"Mw==\",\"created_at\":\"2019-09-02T09:27:17+00:00\",\"updated_at\":\"2019-10-03T12:17:48+00:00\",\"conditions\":[{\"id\":18,\"version\":\"MA==\",\"created_at\":\"2019-10-03T12:17:48+00:00\",\"updated_at\":\"2019-10-03T12:17:48+00:00\"},{\"id\":17,\"version\":\"MA==\",\"created_at\":\"2019-10-03T12:17:48+00:00\",\"updated_at\":\"2019-10-03T12:17:48+00:00\"}],\"photo\":{\"id\":79293,\"location\":\"https:\\/\\/surehub.s3.amazonaws.com\\/user-photos\\/thm\\/23412\\/z70LUtqaHVhlsdfuyHKJH5HDysg5AR6GvQwdAZptCgeZU.jpg\",\"uploading_user_id\":52815,\"version\":\"MA==\",\"created_at\":\"2019-09-02T09:31:07+00:00\",\"updated_at\":\"2019-09-02T09:31:07+00:00\"},\"position\":{\"tag_id\":60456,\"device_id\":318986,\"where\":1,\"since\":\"2019-10-03T10:23:37+00:00\"},\"status\":{\"activity\":{\"tag_id\":60456,\"device_id\":318986,\"where\":1,\"since\":\"2019-10-03T10:23:37+00:00\"}}}";
         SurePetcarePet response = SurePetcareConstants.GSON.fromJson(testReponse, SurePetcarePet.class);
 
-        assertEquals(Integer.valueOf(34675), response.id);
+        assertEquals(Long.valueOf(34675), response.id);
         assertEquals("Cat", response.name);
         assertEquals(Integer.valueOf(0), response.genderId);
 
@@ -105,8 +105,8 @@ public class SurePetcarePetTest {
         assertEquals(BigDecimal.valueOf(3.5), response.weight);
 
         assertEquals("Test Comment", response.comments);
-        assertEquals(Integer.valueOf(87435), response.householdId);
-        assertEquals(Integer.valueOf(23412), response.photoId);
+        assertEquals(Long.valueOf(87435), response.householdId);
+        assertEquals(Long.valueOf(23412), response.photoId);
         assertEquals(SurePetcarePet.PetSpecies.CAT.id, response.speciesId);
         assertEquals(Integer.valueOf(382), response.breedId);
 
@@ -123,7 +123,7 @@ public class SurePetcarePetTest {
         String testReponse = "{\"id\":30622,\"name\":\"Cat\",\"gender\":1,\"date_of_birth\":\"2016-04-01T00:00:00+00:00\",\"weight\":\"6\",\"comments\":\"\",\"household_id\":21005,\"breed_id\":382,\"food_type_id\":1,\"photo_id\":77957,\"species_id\":1,\"tag_id\":24725,\"version\":\"OA==\",\"created_at\":\"2018-12-22T08:59:13+00:00\",\"updated_at\":\"2019-08-26T18:17:38+00:00\",\"photo\":{\"id\":77957,\"location\":\"https://surehub.s3.amazonaws.com/user-photos/thm/22360/1jhp4OtwmNvWXrsT4pWLJhoYOt7Ti9UVm5SjsFoC9Y.jpg          \",\"uploading_user_id\":22360,\"version\":\"MA==\",\"created_at\":\"2019-08-26T18:17:38+00:00\",\"updated_at\":\"2019-08-26T18:17:38+00:00\"},\"position\":{\"tag_id\":24725,\"device_id\":243573,\"where\":2,\"since\":\"2020-05-01T06:01:53+00:00\"},\"status\":{\"activity\":{\"tag_id\":24725,\"device_id\":243573,\"where\":2,\"since\":\"2020-05-01T06:01:53+00:00\"}}}";
         SurePetcarePet response = SurePetcareConstants.GSON.fromJson(testReponse, SurePetcarePet.class);
 
-        assertEquals(Integer.valueOf(30622), response.id);
+        assertEquals(Long.valueOf(30622), response.id);
         assertEquals("Cat", response.name);
         assertEquals(Integer.valueOf(1), response.genderId);
 
@@ -136,8 +136,8 @@ public class SurePetcarePetTest {
         assertEquals(BigDecimal.valueOf(6), response.weight);
 
         assertEquals("", response.comments);
-        assertEquals(Integer.valueOf(21005), response.householdId);
-        assertEquals(Integer.valueOf(77957), response.photoId);
+        assertEquals(Long.valueOf(21005), response.householdId);
+        assertEquals(Long.valueOf(77957), response.photoId);
         assertEquals(SurePetcarePet.PetSpecies.CAT.id, response.speciesId);
         assertEquals(Integer.valueOf(382), response.breedId);
